- List => Collection
- Move receive buffer factory to own file
- DispatchFrom => ConsumeFrom inline with previous rename
